+++ paddle/phi/kernels/gpu/dropout_grad_kernel.cu
@@ -14,9 +14,9 @@
 
 #include "paddle/phi/kernels/dropout_grad_kernel.h"
 
-#include "paddle/fluid/operators/dropout_impl.cu.h"
 #include "paddle/phi/backends/gpu/gpu_context.h"
 #include "paddle/phi/core/kernel_registry.h"
+#include "paddle/phi/kernels/funcs/dropout_impl.cu.h"
 
 namespace phi {
